+++ R/colour-text.R
@@ -3,11 +3,12 @@
 testthat_colours <- list(
   success = green,
   skip = yellow,
+  warning = yellow,
   failure = red,
   error = red
 )
 
-colourise <- function(text, as = c("success", "skip", "failure", "error")) {
+colourise <- function(text, as = c("success", "skip", "warning", "failure", "error")) {
   colour_config <- getOption("testthat.use_colours", TRUE)
   if (!isTRUE(colour_config)) {
     return(text)